ADD : Add Skeleton, Select All button in Filtering CHANGE : supportType_hour to Float

+++ src/main/java/com/example/smstest/domain/support/dto/ModifyRequest.java
@@ -37,6 +37,6 @@ public class ModifyRequest {
 
     private Long supportTypeId;
 
-    private Integer supportTypeHour;
+    private Float supportTypeHour;
 
 }
